update review list when user gives a new rating

+++ app/src/main/java/UtilityClass/CourseRating.java
@@ -8,11 +8,11 @@ import java.util.HashMap;
  */
 
 public class CourseRating {
-    public int assignmentDifficulty, lectureQuality, overallQuality, totalRatings;
+    public long assignmentDifficulty, lectureQuality, overallQuality, totalRatings;
 
     public CourseRating(){}
 
-    public CourseRating(int assignmentDifficulty, int lectureQuality, int overallQuality, int totalRatings){
+    public CourseRating(long assignmentDifficulty, long lectureQuality, long overallQuality, long totalRatings){
         this.assignmentDifficulty = assignmentDifficulty;
         this.lectureQuality = lectureQuality;
         this.overallQuality = overallQuality;
